requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0
requests>=2.28.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
matplotlib>=3.7.0
streamlit>=1.30.0
